fix compilation errors for ESP32-C3
override, uint32_t cast for OSCMessage.add
why NUM_SERVOS is not defined in ServoMotor and ServoManager ?

+++ src/communication/OSCManager.cpp
@@ -169,7 +169,7 @@ void OSCManager::yo(String firmwareVersion)
     answerMsg.add(firmwareVersion.c_str());
     answerMsg.add(wifi->getMAC().c_str());
     answerMsg.add(wifi->getIP().c_str());
-    answerMsg.add(OSC_LISTENING_PORT);
+    answerMsg.add((int32_t)OSC_LISTENING_PORT);
     sendMessage(answerMsg);
 }
